backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-mssql-backend
pyodbc
python-dotenv
coreapi
django-cors-headers
